create new medication + new trip model + refactoring

+++ src/main/java/com/musula/backend/drones/repository/DroneRepository.java
@@ -7,7 +7,7 @@ import com.musula.backend.util.repository.QueryDslPagingAndSortingRepository;
 import java.util.Optional;
 
 
-public interface DroneRepository extends QueryDslPagingAndSortingRepository<Drone, Long, QDrone> {
+public interface DroneRepository extends QueryDslPagingAndSortingRepository<Drone, Integer, QDrone> {
 
     Optional<Drone> findBySerialNumber(String serialNumber);
